+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
+## [2.1.0] - 2018-10-03
+### Changed
+- Fix error on login when user does not have any repository
+- Fix error on repository list or SSH list when list is empty
+
 ## [2.0.0] - 2018-06-16
 ### Added
 - Add SSH keys handling (view, upload with paste or drag-and-drop, delete)